+++ CHANGELOG.md
@@ -1,10 +1,18 @@
+# 0.10.12 Release:
+## New Content:
+- Oath of the Ancients Paladin: Channel Divinity: Nature's Wrath and Channel Divinity: Turn the Faithless
+- Feat: Heavy Armor Master
+## Bug Fixes:
+- Twilight Sanctuary
+## Update Notes:
+- Many macros were adjusted for this update. There may be bugs, please report them on my Github page or Discord server.
+- This is a beta version for the 3.0.3 D&D system. Do not update to this version unless unless you have all the required up to date dependency modules.
+  
 # 0.10.11 Release:
 ## Bug Fixes:
 - Manual Rolls Target Select Dialog
 ## Update Notes:
 - Added more user interface settings.
-- Many macros were adjusted for this update. There may be bugs, please report them on my Github page or Discord server.
-- This is a beta version for the 3.0.3 D&D system. Do not update to this version unless unless you have all the required up to date dependency modules.
   
 # 0.10.10 Release:
 ## New Content: